backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-core
langchain-groq
langgraph
sqlalchemy